LayoutTests/canvas/philip/tests/2d.imageData.create1.basic-expected.txt

< [index] >
2d.imageData.create1.basic
createImageData(imgdata) exists and returns something
References: 2d.imageData.create1.object Bugzilla
Actual output:
Passed
